refactor: 將所有跟金錢相關的型態都改為decimal

+++ Shared/Domain/Block.cs
@@ -35,13 +35,13 @@ public abstract class Block
 
 public class Land: Block
 {
-    private readonly int _price;
+    private readonly decimal _price;
     private int house;
-    public int Price => _price; // 土地購買價格
+    public decimal Price => _price; // 土地購買價格
     // public int UpgradePrice => _price; // 升級價格
     // public int TollFee => _price; // 過路費
     public int House => house;
-    public Land(string id, int price = 1000) : base(id)
+    public Land(string id, decimal price = 1000) : base(id)
     {
         _price = price;
     }
